Fix: Migration

+++ database/migrations/2023_03_13_184851_create_cities_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
     {
         Schema::create('cities', function (Blueprint $table) {
             $table->integer('city_id')->primary();
-            $table->unsignedBigInteger('province_id');
+            $table->integer('province_id');
             $table->foreign('province_id')->references('province_id')->on('provinces');
             $table->string('name')->nullable();
             $table->string('type')->nullable();
