+++ database/migrations/2023_03_13_186641_create_addresses_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->string('label')->nullable();
             $table->string('name')->nullable();
             $table->string('phone')->nullable();
-            $table->unsignedBigInteger('subdistrict_id')->nullable();
+            $table->integer('subdistrict_id')->nullable();
             $table->foreign('subdistrict_id')->references('subdistrict_id')->on('subdistricts');
             $table->text('address')->nullable();
             $table->string('postal_code')->nullable();
